+++ internal/loraserver/loraserver.go
@@ -246,17 +246,27 @@ func handleDataDownReply(ctx Context, rxPacket models.RXPacket, ns models.NodeSe
 
 	// check if there are payloads pending in the queue
 	txPayload, remaining, err := getTXPayloadAndRemainingFromQueue(ctx.RedisPool, ns.DevEUI)
-
-	// errEmptyQueue should not be handled as an error, it just means there
-	// is no queue / the queue is empty
-	if err != nil && err != errEmptyQueue {
+	if err != nil {
 		return err
 	}
 
 	// nothing pending in the queue and no need to ACK RXPacket
-	if rxPacket.PHYPayload.MHDR.MType != lorawan.ConfirmedDataUp && err == errEmptyQueue {
+	if rxPacket.PHYPayload.MHDR.MType != lorawan.ConfirmedDataUp && txPayload == nil {
 		return nil
 	}
+
+	// get TX DR
+	uplinkDR, err := band.GetDataRate(rxPacket.RXInfo.DataRate)
+	if err != nil {
+		return err
+	}
+	// get RX1 frequency
+	rx1Frequency, err := band.GetRX1Frequency(rxPacket.RXInfo.Frequency, uplinkDR)
+	if err != nil {
+		return err
+	}
+	// get RX1 DR
+	rx1DR := band.RX1DROffsetConfiguration[uplinkDR][0] // DR offset is not yet implemented
 
 	phy := lorawan.PHYPayload{
 		MHDR: lorawan.MHDR{
@@ -276,39 +286,60 @@ func handleDataDownReply(ctx Context, rxPacket models.RXPacket, ns models.NodeSe
 	phy.MACPayload = macPL
 
 	// add the payload from the queue
-	if err == nil {
-		macPL.FHDR.FCtrl.FPending = remaining
-
-		if txPayload.Confirmed {
-			phy.MHDR.MType = lorawan.ConfirmedDataDown
-		}
-		macPL.FPort = &txPayload.FPort
-		macPL.FRMPayload = []lorawan.Payload{
-			&lorawan.DataPayload{Bytes: txPayload.Data},
-		}
-		if err := phy.EncryptFRMPayload(ns.AppSKey); err != nil {
-			return fmt.Errorf("encrypt FRMPayload error: %s", err)
-		}
-
-		// remove the payload from the queue when not confirmed
-		if !txPayload.Confirmed {
+	if txPayload != nil {
+		// validate the max payload size
+		if len(txPayload.Data) > band.MACPayloadSizeConfiguration[rx1DR].N {
+			// remove the payload from the queue regarding confirmed or not
 			if err := clearInProcessTXPayload(ctx.RedisPool, ns.DevEUI); err != nil {
 				return err
+			}
+
+			log.WithFields(log.Fields{
+				"dev_eui":             ns.DevEUI,
+				"data_rate":           rx1DR,
+				"frmpayload_size":     len(txPayload.Data),
+				"max_frmpayload_size": band.MACPayloadSizeConfiguration[rx1DR].N,
+			}).Warning("downlink payload max size exceeded")
+			err = ctx.Application.SendNotification(ns.DevEUI, ns.AppEUI, models.ErrorNotificationType, models.ErrorNotification{
+				Reference: txPayload.Reference,
+				DevEUI:    ns.DevEUI,
+				Message:   fmt.Sprintf("downlink payload max size exceeded (dr: %d, allowed: %d, got: %d)", rx1DR, band.MACPayloadSizeConfiguration[rx1DR].N, len(txPayload.Data)),
+			})
+			if err != nil {
+				return err
+			}
+		} else {
+			// remove the payload from the queue when not confirmed
+			if !txPayload.Confirmed {
+				if err := clearInProcessTXPayload(ctx.RedisPool, ns.DevEUI); err != nil {
+					return err
+				}
+			}
+
+			macPL.FHDR.FCtrl.FPending = remaining
+			if txPayload.Confirmed {
+				phy.MHDR.MType = lorawan.ConfirmedDataDown
+			}
+			macPL.FPort = &txPayload.FPort
+			macPL.FRMPayload = []lorawan.Payload{
+				&lorawan.DataPayload{Bytes: txPayload.Data},
 			}
 		}
 	}
 
-	if err := phy.SetMIC(ns.NwkSKey); err != nil {
-		return fmt.Errorf("set MIC error: %s", err)
+	// when the payload did not pass the validation and there is no ACK set,
+	// there is nothing to send
+	if !macPL.FHDR.FCtrl.ACK && len(macPL.FRMPayload) == 0 {
+		return nil
 	}
 
-	dr, err := band.GetDataRate(rxPacket.RXInfo.DataRate)
-	if err != nil {
-		return err
+	// if there is no payload set, encrypt will just do nothing
+	if err := phy.EncryptFRMPayload(ns.AppSKey); err != nil {
+		return fmt.Errorf("encrypt FRMPayload error: %s", err)
 	}
-	rx1Frequency, err := band.GetRX1Frequency(rxPacket.RXInfo.Frequency, dr)
-	if err != nil {
-		return err
+
+	if err := phy.SetMIC(ns.NwkSKey); err != nil {
+		return fmt.Errorf("set MIC error: %s", err)
 	}
 
 	txPacket := models.TXPacket{
@@ -317,7 +348,7 @@ func handleDataDownReply(ctx Context, rxPacket models.RXPacket, ns models.NodeSe
 			Timestamp: rxPacket.RXInfo.Timestamp + uint32(band.ReceiveDelay1/time.Microsecond),
 			Frequency: rx1Frequency,
 			Power:     band.DefaultTXPower,
-			DataRate:  band.DataRateConfiguration[band.RX1DROffsetConfiguration[dr][0]], // currently offset is not configurable
+			DataRate:  band.DataRateConfiguration[rx1DR],
 			CodeRate:  rxPacket.RXInfo.CodeRate,
 		},
 		PHYPayload: phy,