+++ models/notifications.go
@@ -14,6 +14,7 @@ type NotificationType int
 // Available notification types.
 const (
 	JoinNotificationType = iota
+	ErrorNotificationType
 )
 
 // JoinNotification defines the payload sent to the application on
@@ -22,4 +23,12 @@ type JoinNotification struct {
 	DevAddr lorawan.DevAddr `json:"devAddr"`
 	DevEUI  lorawan.EUI64   `json:"devEUI"`
 	Time    time.Time       `json:"time"`
+}
+
+// ErrorNotification defines the payload sent to the application
+// on an error event.
+type ErrorNotification struct {
+	Reference string        `json:"reference"` // refers to the given reference by the application
+	DevEUI    lorawan.EUI64 `json:"devEUI"`
+	Message   string        `json:"message"`
 }